fix XChangeProperty call in XdndEnable: format 32 should use long; fixes crash on Solaris 10 sparc

+++ util/xdnd.c
@@ -142,8 +142,10 @@ void XdndInit(
     dropData = dropCBData;
 }
 
+
+
 void XdndEnable(Widget w) {
-    int version = 4;
+    long version = 4;
     XChangeProperty(
             XtDisplay(w),
             XtWindow(w),
